refactor(hooks): centralize vendor detection in .agents/hooks/core (P6)
Six hook handlers each carried their own copy of inferVendorFromScriptPath
(x4), getProjectDir (x6), and getHookDir — which had drifted to different
vendor subsets (removing one vendor meant editing all of them). Extract them
into a shared vendor-detect.ts (canonical union) that every handler imports;
inferVendorFromScriptPath now takes the caller's import.meta.filename.

detectVendor stays per-handler (the event-name->vendor mapping is genuinely
hook-kind-specific: prompt/tool/stop) but delegates the shared script-path
inference. vendor-wiring.test imports getHookDir from the shared module.

Net: getProjectDir 6->1, inferVendorFromScriptPath 4->1. A vendor add/remove
now touches one detection module instead of six.
