+++ cli/commands/hook/vendor-wiring.test.ts
@@ -27,8 +27,8 @@ import { readdirSync, readFileSync } from "node:fs";
 import { join } from "node:path";
 import { describe, expect, it } from "vitest";
 import { VENDORS } from "../../../.agents/hooks/core/constants.js";
-import { getHookDir } from "../../../.agents/hooks/core/test-filter.js";
 import type { Vendor } from "../../../.agents/hooks/core/types.js";
+import { getHookDir } from "../../../.agents/hooks/core/vendor-detect.js";
 import { nativeEventToKind } from "./adapters.js";
 import { VARIANT_ROUTES } from "./dispatch.js";
 import { PROBE_VENDORS } from "./probe.js";
